add comments and .gitignore file

+++ T1-5.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-int Factorial(int); // prototype
+int Factorial(int); // function prototype
 
 int main(){
     int n ;
@@ -11,11 +11,11 @@ int main(){
     std::cin >> n;
 
     for (int i=1; i <= n; i++)
-        sum += static_cast<double>(1.0) / Factorial(i);
+        sum += 1.0 / Factorial(i);
     std::cout << "sum : " << sum << std::endl;
 }
 
-int Factorial(int value) {
+int Factorial(int value) { // recursive function 
     if ( 1 == value )
         return 1;
     else
